Add trigger results to session triggers for ease of removal and fix tutorials not removing themselves

+++ src/main/java/me/neoblade298/neorogue/equipment/weapons/RighteousLance.java
@@ -1,6 +1,4 @@
 package me.neoblade298.neorogue.equipment.weapons;
-import me.neoblade298.neorogue.equipment.SessionEquipment;
-
 import java.util.LinkedList;
 
 import org.bukkit.Location;
@@ -17,6 +15,7 @@ import me.neoblade298.neorogue.DescUtil;
 import me.neoblade298.neorogue.equipment.Equipment;
 import me.neoblade298.neorogue.equipment.EquipmentProperties;
 import me.neoblade298.neorogue.equipment.Rarity;
+import me.neoblade298.neorogue.equipment.SessionEquipment;
 import me.neoblade298.neorogue.player.inventory.GlossaryTag;
 import me.neoblade298.neorogue.session.fight.DamageMeta;
 import me.neoblade298.neorogue.session.fight.DamageSlice;
@@ -38,7 +37,7 @@ public class RighteousLance extends Equipment {
 	private static final TargetProperties regHit = TargetProperties.line(4, 1, TargetType.ENEMY),
 			lanceHit = TargetProperties.line(6, 1, TargetType.ENEMY);
 	private int sanct;
-	private double dmgMult;
+	private int dmgMult;
 
 	public RighteousLance(boolean isUpgraded) {
 		super(
